Update to version 1.7.1, included in ldns.
Drill is now included in ldns, and the configure machinery has been
re-done, so this now needs to build ldns locally in addition to drill,
but this still only installs drill.  Make the minimal adjustment
so that this builds and installs drill.  Now also works on systems
with OpenSSL 1.1, so remove the BROKEN setting.