pugixml: update to 1.11.4.
1.11.4

This patch reintroduces pugixml target for CMake when find_package
is used, which was accidentally removed from v1.11.  It's recommended
that pugixml::pugixml target is used instead, but this patch restores
pugixml target for existing applications / libraries that depend
on it.

The previous patch, v1.11.3, did not handle repeated attempts to
call find_package; this patch fixes that particular corner case.

1.11.3

The previous patch, v1.11.2, was not compatible with CMake 3.10;
this patch restores compatibility with earlier CMake versions when
using imported targets.

1.11.2

The previous patch, v1.11.1, was only compatible with CMake 3.18
and up; this patch restores compatibility with earlier CMake versions
when using imported targets.