papirus-icon-theme: Switch to dynamic PLIST.
The static PLIST was incorrect, listing a number of directories instead of the
files contained within them.  For some reason the check-files infrastructure
does not complain about this on NetBSD, and was only noticed on SunOS.

Fixing the PLIST to correctly register the 750,000+ files that this package
installs would bloat it to 44MB and be painful to maintain, so while dynamic
PLISTs are generally to be avoided, in this situation it's the least worst.

Ok nia.